refactor: use environment variable for API base URL

+++ src/lib/axios.js
@@ -2,7 +2,7 @@
 import axios from "axios";
 
 const api = axios.create({
-  baseURL: "http://localhost:8000/api",
+  baseURL: process.env.NEXT_PUBLIC_API_URL,
   headers: {
     "Content-Type": "application/json",
   },
@@ -21,26 +21,12 @@ api.interceptors.request.use((config) => {
 });
 
 api.interceptors.response.use(
-  (res) => res,
-  (err) => {
+  (response) => response,
+  (error) => {
     if (
-      err.response?.status === 401 &&
+      error.response?.status === 401 &&
       typeof window !== "undefined"
     ) {
-      localStorage.removeItem("token");
-      window.location.href = "/login";
-    }
-
-    return Promise.reject(err);
-  }
-);
-
-api.interceptors.response.use(
-  response => response,
-  error => {
-
-    if (error.response?.status === 401) {
-
       localStorage.clear();
 
       document.cookie =
